requirements.txt
paho-mqtt==1.6.1
Django==4.1.6
git+https://github.com/ethereum/web3.py.git
py-solc-x==1.1.1
requests==2.28.2
